tidy(proxy): Split file, move to create utils
